Acertei detalhes em perguntas, tarefas e respostas

+++ api/modules/colecoes/ColecaoPerguntaEmBDR.php
@@ -16,15 +16,18 @@ class ColecaoPerguntaEmBDR implements ColecaoPergunta {
 
 	function adicionar(&$obj) {
 		try {	
-			DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+			if($this->validarPergunta($obj)){
+				DB::statement('SET FOREIGN_KEY_CHECKS=0;');
 			
-			$id = DB::table(self::TABELA)->insertGetId(['pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId()]);
-
-			DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-
-			$obj->setId($id);
-
-			return $obj;
+				$id = DB::table(self::TABELA)->insertGetId(['pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId()]);
+	
+				DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+	
+				$obj->setId($id);
+	
+				return $obj;
+			}
+			
 		}
 		catch (\Exception $e)
 		{
@@ -37,9 +40,11 @@ class ColecaoPerguntaEmBDR implements ColecaoPergunta {
 			DB::statement('SET FOREIGN_KEY_CHECKS=0;');
 
 			foreach ($objs as $key => $obj) {
-				$id = DB::table(self::TABELA)->insertGetId([ 'pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId()]);
-				$obj->setId($id);
-				$objs[$key] = $obj;
+				if($this->validarPergunta($obj)){
+					$id = DB::table(self::TABELA)->insertGetId([ 'pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId()]);
+					$obj->setId($id);
+					$objs[$key] = $obj;
+				}
 			}
 
 			DB::statement('SET FOREIGN_KEY_CHECKS=1;');
@@ -54,10 +59,13 @@ class ColecaoPerguntaEmBDR implements ColecaoPergunta {
 
 	function remover($id, $idTarefa) {
 		try {	
-			DB::statement('SET FOREIGN_KEY_CHECKS=0;');
-			$removido = DB::table(self::TABELA)->where('id', $id)->where('tarefa_id', $idTarefa)->delete();
-			DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-			return $removido;
+			if($this->validarRemocaoPergunta($id)){
+
+				DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+				$removido = DB::table(self::TABELA)->where('id', $id)->where('tarefa_id', $idTarefa)->delete();
+				DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+				return $removido;
+			}
 		}
 		catch (\Exception $e)
 		{
@@ -67,14 +75,15 @@ class ColecaoPerguntaEmBDR implements ColecaoPergunta {
 
 	function atualizar(&$obj) {
 		try {
+			if($this->validarPergunta($obj)){
 			
-			DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+				DB::statement('SET FOREIGN_KEY_CHECKS=0;');
 
-			DB::table(self::TABELA)->where('id', $obj->getId())->update([ 'pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId(),'resposta_id' => $obj->getResposta()->getId()
-			]);
+				DB::table(self::TABELA)->where('id', $obj->getId())->update([ 'pergunta' => $obj->getPergunta(), 'tarefa_id' => $obj->getTarefa()->getId()]);
 
-			DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-			return $obj;
+				DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+				return $obj;
+			}
 		}
 		catch (\Exception $e)
 		{
@@ -177,8 +186,34 @@ class ColecaoPerguntaEmBDR implements ColecaoPergunta {
 		return $pergunta;
 	}	
 
-    function contagem() {
-		return DB::table(self::TABELA)->count();
+    function contagem($tarefaId = 0) {
+		return ($tarefaId > 0) ? DB::table(self::TABELA)->where('tarefa_id', $tarefaId)->count() : DB::table(self::TABELA)->count();
+	}
+
+	private function validarPergunta(&$obj) {
+		if(!is_string($obj->getPergunta())) throw new ColecaoException('Valor inválido para titulo.');
+
+		if(strlen($obj->getPergunta()) <= 2 && strlen($obj->getTitulo()) > 255) throw new ColecaoException('A pergunta deve conter no mínimo 2 e no máximo 255.');
+		
+		$quantidade = DB::table(self::TABELA)->whereRaw('pergunta like  "%'. $obj->getPergunta() . '%"')->where('tarefa_id', $obj->getTarefa()->getId())->where(self::TABELA . '.id', '<>', $obj->getId())->count();
+		
+		if($quantidade > 0){
+			throw new ColecaoException('Já exite uma tarefa cadastrada com esse título.');
+		}
+
+		return true;
+	}
+
+	private function validarRemocaoPergunta($id){
+		$quantidade = DB::table(ColecaoRespostaEmBDR::TABELA)->where('pergunta_id', $id)->count();
+
+		if($quantidade > 0)throw new ColecaoException('Não foi possível excluir a pergunta por que ela possui respostas relacionadas a ela. Exclua todas as respostas relacionadas e tente novamente.');
+
+		$quantidade = DB::table(ColecaoFormularioRespondidoEmBDR::TABELA)->where('TABELA_RELACIONAL', $id)->count();
+
+		if($quantidade > 0)throw new ColecaoException('Não é possível excluir uma pergunta já respondida.');
+
+		return true;
 	}
 }
 ?>
